refactored out distribution from particle rep
